Further work on photosynthesis scheme. The code previously computed Anet by leaf layer, and then averaged over leaf layers, and finally multiplied by an effective total leaf area. That is: sum(Anet) * sum(leaf_area)
The code is now intended to compute the sum(Anet * leaf_area).

I further changed the calculation of effective leaf area. There seemed to be
a simple factor-of-two. I also changed the efficiency of mature leaves so that
they are fully productive.

One confusing aspect remaining is the treatment of Rday. This seemed to have been always zero?