Bug fixes to CalcLAI and Fluxh. Added variables t_TotLayerVox (total number of voxels in a given crown layer) and t_OccupyLayerVox (number of voxels occupied by this stem in the layer).
However, I noted that t_leafarea is inconsistent with LAI3D. This is due to rounding.
t_leafarea is computed based on the theoretical canopy volume (Crown_Radius**2 * Crown_Depth).
Voxels filled are either filled or not, based on their proximity to the stem. When crowns are
small, these rounding errors can make LAI3D (summed over columns) differ from t_leafarea by
a factor of 2. Need to deal with this next.